% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pass2.R
\name{get_pass2}
\alias{get_pass2}
\title{Get a password from an environment variable, keychain or prompt}
\usage{
get_pass2(user, password = NULL, service, keyring = NULL)
}
\arguments{
\item{user}{\verb{[character(1)]} giving the username to authenticate with, for example \code{"jane"}.
Also used to retrieve from and store the password in \code{keyring}.}

\item{password}{\verb{[character(1)]} giving the password.
Use this to substitute in an environment variable such as \code{Sys.getenv("SECRET")}.
Recommended only when the environment variable is declared in an encrypted form, and when it is redacted from the logs.
\strong{Do not expose your passwords in scripts or the R console.}
Defaults to \code{NULL}, in which case the \code{keyring} is used.}

\item{service}{\verb{[character(1)]} giving the service to authenticate against with \code{user} and \code{password}, for example \code{"foo-service.com"}.
Also used to retrieve from and store the password in \code{keyring}.
If a longer URL is provided, the domain is extracted using \code{\link[urltools:domain]{urltools::domain()}}, which is the conventional name for storing secrets (for example, \code{"foo-service.com"}, not \code{"https://foo-service.com/api"}).}

\item{keyring}{For systems that support multiple keyrings, specify
the name of the keyring to use here. If \code{NULL}, then the default
keyring is used. See also \code{\link[keyring:has_keyring_support]{has_keyring_support()}}.}
}
\description{
Gets a password, in this order, from:
}
\details{
\enumerate{
\item \code{password}, unless it is \code{NULL} or \code{""} as for an unset variable.
\item a password matching \code{user} and \code{service} in the system's keychain \code{keyring}.
\item an interactive prompt.
}

Otherwise, an error is thrown.
}
