% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate2.R
\name{authenticate2}
\alias{authenticate2}
\title{Use http authentication from environment variable, keychain or prompt
Wraps \code{\link[httr:authenticate]{httr::authenticate()}} for convenience.
See \code{\link[=get_pass2]{get_pass2()}} for details.}
\usage{
authenticate2(user, type = "basic", service, keyring = NULL, env_var = NULL)
}
\arguments{
\item{user}{user name}

\item{type}{type of HTTP authentication.  Should be one of the following
types supported by Curl: basic, digest, digest_ie, gssnegotiate,
ntlm, any.  It defaults to "basic", the most common type.}

\item{service}{\verb{[character(1)]} giving the service to authenticate against, such as \code{"www.foo.com"}.
Used for retrieving from, and storing the secret to they \code{keyring}.
If a longer URL is provided, the domain is extracted using \code{\link[urltools:domain]{urltools::domain()}}, which is the conventional name for storing secrets (for example, \code{"www.fooservice.com"}.).}

\item{keyring}{For systems that support multiple keyrings, specify
the name of the keyring to use here. If \code{NULL}, then the default
keyring is used. See also \code{\link[keyring:has_keyring_support]{has_keyring_support()}}.}

\item{env_var}{\verb{[character(1)]} giving the key (name) to the environment variable under which the password is stored.
Recommended only when the environment variable is declared in an encrypted form, and when it is redacted from the logs.
Use with extreme caution.
Defaults to \code{NULL}, in which case the password is prompted.}
}
\description{
Use http authentication from environment variable, keychain or prompt
Wraps \code{\link[httr:authenticate]{httr::authenticate()}} for convenience.
See \code{\link[=get_pass2]{get_pass2()}} for details.
}
\details{
Key and token-based authentication is more secure and convenient; use basic authentication (BA, aka. username and password) only if unavoidable and at your own risk.
Read the {httr} vignette on \href{https://httr.r-lib.org/articles/secrets.html}{managing secrets} for details.
}
