% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pass2.R
\name{get_pass2}
\alias{get_pass2}
\title{Get a password from an environment variable, keychain or prompt}
\usage{
get_pass2(user, service, keyring = NULL, env_var = NULL)
}
\arguments{
\item{user}{user name}

\item{service}{\verb{[character(1)]} giving the service to authenticate against, such as \code{"www.foo.com"}.
Used for retrieving from, and storing the secret to they \code{keyring}.
If a longer URL is provided, the domain is extracted using \code{\link[urltools:domain]{urltools::domain()}}, which is the conventional name for storing secrets (for example, \code{"www.fooservice.com"}.).}

\item{keyring}{For systems that support multiple keyrings, specify
the name of the keyring to use here. If \code{NULL}, then the default
keyring is used. See also \code{\link[keyring:has_keyring_support]{has_keyring_support()}}.}

\item{env_var}{\verb{[character(1)]} giving the key (name) to the environment variable under which the password is stored.
Recommended only when the environment variable is declared in an encrypted form, and when it is redacted from the logs.
Use with extreme caution.
Defaults to \code{NULL}, in which case the password is prompted.}
}
\description{
If \code{env_var} is found, it's value is used as a password.
If \code{env_var} is unset (\code{""}) or if \code{env_var} is \code{NULL}, a password matching \code{user} and \code{service} in \code{keyring} is used.
If no password in the \code{keyring} is found, the keychain unavailable or locked but the session is interactive, a prompt is opened.
Otherwise, an error is thrown.
}
